fix: use the node modules paths for every resolve request
This should force Metro always to resolve libraries using the explicit `nodeModulesPaths` we provide, and in that specific order too. This should allow you to use multiple react/react-dom/react-native versions without breaking them.

Note, this is highly experimental, and we need to check with the Metro team if this is safe to do.